[project @ 2001-10-03 13:57:33 by simonpj] ------------------------------------------- 	Don't bomb out on inlining big constructors 	-------------------------------------------
The unfolder bombs out if the expression get too big; no point in
computing its exact size when it's vast.  But for *constructors*
applied to a lot of args, it's worth keeping going, because we
get a big *disount* too
	$fFooInt = :CFoo a1 a2 a3 ... a50

We want to keep the inlining for $fFooInt in interface files, so
that importing guys can do the selection.


Solution: only bomb out when size-discount gets too big.