[project @ 2001-10-17 11:05:36 by simonpj] ------------------------------------------- 	Desugar bindings into Rec groups more often 	-------------------------------------------
In rather obscure cases (involving functional dependencies)
it is possible to get an AbsBinds [] [] (no tyvars, no dicts)
which nevertheless has some "dictionary bindings".  These
come out of the typechecker in non-dependency order, so we
need to Rec them just in case.

It turns out to be a bit awkward.  The smallest fix is
to make dsLet always make a Rec; brutal but correct.