[project @ 2001-10-25 14:30:43 by simonpj] -------------------------------------------------------   Correct an error in the handling of implicit parameters   -------------------------------------------------------
	MERGE WITH STABLE BRANCH UNLESS HARD TO DO

Mark Shields discovered a bug in the way that implicit parameters
are dealt with by the type checker.  It's all a bit subtle, and
is extensively documented in TcSimplify.lhs.

This commit makes the code both simpler and more correct.  It subtly
changes the way in which type signatures are treated, but not in a way
anyone would notice: see notes with "Question 2: type signatures"
in TcSimplify.lhs.