[project @ 2001-10-31 15:22:53 by simonpj] ------------------------------------------ 	Improved handling of scoped type variables 	------------------------------------------
The main effect of this commit is to allow scoped type variables
in pattern bindings, thus

	(x::a, y::b) = e

This was illegal, but now it's ok.  a and b have the same scope
as x and y.


On the way I beefed up the info inside a type variable
(TcType.TyVarDetails; c.f. IdInfo.GlobalIdDetails) which
helps to improve error messages. Hence the wide ranging changes.
Pity about the extra loop from Var to TcType, but can't be helped.