[project @ 2001-11-01 10:33:58 by simonpj] ---------------------------------- 	Fix a bug in Simplify.mkDupableAlt 	----------------------------------
	[This is the HEAD commit; I've fixed
	 the branch separately.]

This fixes a funResultTy panic that Koen encountered.

The reason was that the simplifier was doing a
case-of-case where the result had a polymorphic type.
This in turn showed up because of a newtype (now
transparent) with a forall inside it.

The fix is very easy; can't think how I got it wrong
in the first place.