[project @ 2001-11-01 13:20:05 by simonpj] --------------------------------------- 	Fix a unboxed-binding bug in SpecConstr 	---------------------------------------
	[HEAD only]

This fixes a rather obscure bug in the constructor
specialiser discovered by Ralf Hinze.  It was
generating a specialised version of the function
with no arguments --- and the function returned an
unboxed type.

Solution: same as for worker-wrapper; add a dummy
argument.

Several files are affected because I added
CoreUtils.mkPiTypes, as a useful helper function.