[project @ 2001-11-09 16:41:15 by simonpj] --------------------------------------- 	Fix an obscure overlapping-instance bug 	---------------------------------------
	MERGE TO STABLE BRANCH

When searching for instances, we used bale out if the type
we seek could be instantiated to match the instance (because
it might be so instantiated later, in which case we don't
want to miss the opportunity).

The bug was that we used *matching* whereas we should use
*unification*.

Comments in the file InstEnv.