[project @ 2001-11-19 14:23:52 by simonpj] -------------------------------------- 	Yet another cut at the DmdAnal domains 	--------------------------------------
This version of the domain for demand analysis was developed
in discussion with Peter Sestoft, so I think it might at last
be more or less right!

Our idea is mentally to separate
	strictness analysis
from
	absence and boxity analysis

Then we combine them back into a single domain.  The latter
is all you see in the compiler (the Demand type, as before)
but we understand it better now.