[project @ 2002-01-22 13:35:36 by simonmar] Attempt to fix the problems with missing instances once more.
The current problem is that in the case where a ModDetails is being
constructed from its interface (in compilation manager modes) we
weren't getting any instances because the instances are gotten from
the [InstInfo] returned from tcInstDecls1, which only contains
*source* instance declarations.  Fix: return a list of DFuns defined
in the current module from tcInstDecls1, to be plugged into the
ModDetails later.

Also: revert the previous change to the isLocalThing predicate,
because now we really want to know which dfuns come from the current
module.  The comment about the iface_dfuns containing only package and
local instances is incorrect in batch-compile mode, because we also
demand-load stuff from home package interfaces, so I deleted this
comment and fixed up some of the other commentary.