[project @ 2002-01-25 16:46:53 by simonmar] In GHCi, if we are currently using a compiled version of a module and the user compiles a new version of the module, allow the new version to be linked in during a :reload.  (as suggested by Koen Claessen).
We can't go all the way and allow a newly compiled module to replace
an existing interpreted version, because the version numbers in the
interface file will be out-of-sync with our internal copy of the
interface.  To link in a newly compiled version of an interpreted
module, you still have to do :load.